Factor out the test setup for plain command handler syntax.
